+++ src/main/java/pl/plajer/villagedefense/database/FileStats.java
@@ -24,6 +24,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 
+import pl.plajer.villagedefense.ConfigPreferences;
 import pl.plajer.villagedefense.Main;
 import pl.plajer.villagedefense.api.StatsStorage;
 import pl.plajer.villagedefense.arena.ArenaRegistry;
@@ -69,10 +70,10 @@ public class FileStats {
 
   public void loadStatsForPlayersOnline() {
     for (final Player player : plugin.getServer().getOnlinePlayers()) {
-      if (plugin.isBungeeActivated()) {
+      if (plugin.getConfigPreferences().getOption(ConfigPreferences.Option.BUNGEE_ENABLED)) {
         ArenaRegistry.getArenas().get(0).teleportToLobby(player);
       }
-      if (!plugin.isDatabaseActivated()) {
+      if (!plugin.getConfigPreferences().getOption(ConfigPreferences.Option.DATABASE_ENABLED)) {
         for (StatsStorage.StatisticType s : StatsStorage.StatisticType.values()) {
           loadStat(player, s);
         }